requirements.txt
aiofiles==23.2.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.1.2
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
fastapi==0.109.2
h11==0.16.0
idna==3.10
motor==3.3.2
packaging==25.0
passlib==1.7.4
pefile==2023.2.7
pillow==10.2.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pymongo==4.6.1
python-dotenv==1.0.1
python-jose==3.3.0
python-jose[cryptography]==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
pywin32-ctypes==0.2.3
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.27.1
sqlalchemy==2.0.27
numpy==1.26.4
opencv-python==4.9.0.80
pydicom==2.4.4
scikit-image==0.22.0
tensorflow==2.15.0
tensorflow-hub==0.15.0
